WebView: Separate construction and loading
WebView had two ctors, one for when the page should be loaded directly
from a URL and one when the URL was to be combined with a search query
first. Both had lots of common code that was clumped together with
loading code that doesn't belong in a ctor.

This change makes it so there's a single ctor and loading has to be
explicitly initiated by calling load(). Still far from pretty, but
removes lots of redundancy.