Remove "Ruby(subscription) on/off" option
When this option was off, all <sub> HTML tags would be stripped, but not
their contents, thereby blending them into the text. This seems pretty
useless, so the option likely won't be missed.

Removing it gets rid of another #ifdef and a bool that was passed all
over the place.