Windows: Add build option to use UTF-8 for EB path names
On Windows, the EB library can only access paths which are representable
in the current code page. For instance, opening a dictionary in a path
containing "辞書" works on a Japanese Windows, but not a European one.

This change lets qolibri convert path names to and from UTF-8 at the EB
API boundary. With an appropriately patched version of EB, this allows
access to the full Unicode range of path names. With an unmodified EB,
this reduces compatibility to 7-bit ASCII.