Model: Remove redundant scanClipboard flag
This information is now kept solely in the "checked" property of
MainWindow's scan clipboard action object.

This has the user-visible side effect of disabling the "Scan clipboard"
option on the next run, because the flag is now persisted for the bogus
"EpwingViewer" application as opposed to the bogus "option" application.