OptionDialog: Add "Clipboard" tab (where supported)
X11: Allow watching either the clipboard or the primary selection. For
the latter, add a configurable delay between 0 and 5 seconds with the
previously hard-coded 300ms as a default. Also add option to disable
raising the window on clipboard events, as this causes annoying
focus stealing, at least on i3.

macOS: Experimental: Allow watching either the clipboard or the mythical
"find buffer". I couldn't find any official documentation on this
feature, apart from deprecated API docs. According to folklore it should
be mapped to ⌘E, but on my Mac this just opens the finder, regardless of
app. If the legends are indeed true, this could be useful for some, so
let's support it for completeness.

All platforms: Ignore clipboard events while modal dialogs are open.